d/libgtk2.0-0.symbols: Ignore removal of g_cclosure_marshal_ symbols
These symbols are all canonically provided by GObject, part of the
libglib2.0-0 package, on which GTK depends. They were accidentally
included in the GTK 2.24.32 ABI, but if they are removed, dependent
programs will find the copies in GObject, which are completely
compatible (it has the same content and was generated by the same
script).