d/tests/1065493-futureproofing: Add a test for recurrence of #1065493
If at some point in the future we have another transition as extensive
as time64, then libgtk2.0-0t64 could conceivably be replaced by some
other package, which I have modelled here as libgtk2.0-0xyz. If that
happens, we need to avoid deletion of immmodules.cache, otherwise
we will have another bug similar to #1065493.

This test-case depends on several implementation details of dpkg-repack
and libgtk2.0-0t64, so it might need to be adjusted in the future. As
a result, I have marked it as flaky, so that failures in the official
autopkgtest environment will not be considered a release-critical bug
that stalls migration and requires immediate intervention by maintainers.