d/libgtk2.0-0t64.postrm: Avoid recurrence of #1065493 in the future
If at some point in the future we have another transition as extensive
as time64, then libgtk2.0-0t64 could conceivably be replaced by some
other package, which I have modelled here as libgtk2.0-0xyz. If that
happens, we need to avoid deletion of immodules.cache, otherwise we
will have another bug similar to #1065493.

This implementation is based on the assumption that third-party input
method modules for GTK 2 will depend on GTK 2, therefore we should not
need to clean up the IM modules cache unless/until we reach the point
of having no IM modules installed.